Separate endpoint tests
